Reclaim 402 status code
402 was used previously in some cases where account confirmation was
required (it seems that the `VerificationRequired` exception here was a
misnomer), but confirmation is a concept that we're trying to deprecate
anyway as indicated in heroku/api#750.

Remove specialized use of 402, and make it a generic "Payment Required"
exception once again so that we can re-use it to signal account
deliquency.